requirements.txt
pyinstaller>=6.11.0,<6.12.0
PyQt6>=6.8.0,<6.9.0

# Extract images from PDF
PyMuPDF==1.25.2

# Wrapper for Tesseract OCR
pytesseract==0.3.13

# PIL
pillow==11.1.0

pypdf==5.2.0
